[Improvement-3369][api] Introduce executor and datasource service interface for clear code
